Implemented Coka's changes, Thanks Coka!
