+++ sites/skeleton.dev/src/routes/blog/[slug]/+page.svelte
@@ -19,12 +19,12 @@
 		// Element Page
 		elemPage = document.querySelector('#page');
 		// CodeBlock Highlight
-		document.querySelectorAll('pre code').forEach((elem: Element) => {
+		document.querySelectorAll('pre code').forEach((elem) => {
 			if (!(elem instanceof HTMLElement)) return;
 			hljs.highlightElement(elem);
 		});
 		// Table
-		document.querySelectorAll('table').forEach((elem: Element) => {
+		document.querySelectorAll('table').forEach((elem) => {
 			elem.classList.add('table');
 		});
 	});
